Moved bigint package to kerl
